+++ sparse_strips/vello_hybrid/examples/wgpu_webgl/src/lib.rs
@@ -226,6 +226,8 @@ impl AppState {
                 &render_size,
                 &surface_texture_view,
                 &vello_hybrid::TextureBindings::new(),
+                vello_hybrid::ClearSettings::default(),
+                vello_hybrid::RenderRegion::Full,
             )
             .unwrap();
 
@@ -657,6 +659,8 @@ pub async fn render_scene(scene: Scene, width: u16, height: u16) {
             &render_size,
             &surface_texture_view,
             &vello_hybrid::TextureBindings::new(),
+            vello_hybrid::ClearSettings::default(),
+            vello_hybrid::RenderRegion::Full,
         )
         .unwrap();