+++ sparse_strips/vello_hybrid/examples/winit/src/main.rs
@@ -403,6 +403,8 @@ impl ApplicationHandler for App<'_> {
                         &render_size,
                         &texture_view,
                         &texture_bindings,
+                        vello_hybrid::ClearSettings::default(),
+                        vello_hybrid::RenderRegion::Full,
                     )
                     .unwrap();